+++ Source/RuntimeAudioImporter/RuntimeAudioImporter.Build.cs
@@ -6,11 +6,6 @@ using UnrealBuildTool;
 
 public class RuntimeAudioImporter : ModuleRules
 {
-	private string ThirdPartyPath
-	{
-		get { return Path.GetFullPath(Path.Combine(ModuleDirectory, "../ThirdParty/")); }
-	}
-
 	public RuntimeAudioImporter(ReadOnlyTargetRules Target) : base(Target)
 	{
 		PCHUsage = ModuleRules.PCHUsageMode.UseExplicitOrSharedPCHs;
@@ -29,8 +24,6 @@ public class RuntimeAudioImporter : ModuleRules
 				"DR_FLAC_IMPLEMENTATION=1"
 			}
 		);
-
-		PrivateIncludePaths.Add(ThirdPartyPath);
 
 		PrivateDependencyModuleNames.AddRange(
 			new string[]
